Add env var to enable persisting of working directory and files after acceptance test has completed (#18)

+++ internal/plugintest/helper.go
@@ -95,6 +95,11 @@ func (h *Helper) Close() error {
 			return err
 		}
 	}
+
+	if v := os.Getenv(EnvTfAccPersistWorkingDir); v != "" {
+		return nil
+	}
+
 	return os.RemoveAll(h.baseDir)
 }
